factor out optimization functions into framework module
